% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ngrams.R
\name{add_1grams}
\alias{add_1grams}
\title{Add 1-grams}
\usage{
add_1grams(ngram, u)
}
\arguments{
\item{ngram}{a single n-gram.}

\item{u}{\code{integer}, \code{numeric} or \code{character} vector of all possible 
unigrams.}
}
\value{
vector of n-grams (where \code{n} is equal to the \code{n} of the input plus one) 
with position information.
}
\description{
Builds (n+1)-grams from n-grams.
}
\details{
n-grams are built by pasting every possible unigram in the every possible free 
position.
}
\examples{
add_1grams("1_2.3.4_3.0", 1L:5)
}

