% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{binarize}
\alias{binarize}
\title{Binarize}
\usage{
binarize(x)
}
\arguments{
\item{x}{\code{matrix} or \code{\link[slam]{simple_triplet_matrix}}.}
}
\value{
a \code{matrix} or \code{\link[slam]{simple_triplet_matrix}} 
(depending on the input).
}
\description{
Binarizes a matrix.
}
