% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{biogram-package}
\alias{biogram}
\alias{biogram-package}
\title{biogram - analysis of n-grams from biological sequences}
\description{
biogram specialises in analysis of n-grams (k-tuples) from nucleotide or
protein sequences.
}
\author{
Michal Burdukiewicz
}

