% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build_ngrams.R
\name{build_ngrams}
\alias{build_ngrams}
\title{Build n-grams}
\usage{
build_ngrams(ngrams)
}
\arguments{
\item{ngrams}{a vector of positioned n-grams (as created by \code{\link{count_ngrams}}).}
}
\value{
a vector of n-grams (where \code{n} is equal to the \code{n} of the input plus one)
with position information.
}
\description{
Builds longer n-grams from shorter n-grams.
}
\details{
n-grams are build iteratively by pasting existing n-grams with unigrams extracted
from them.
}
\note{
All n-grams must have the same length (\code{n}).
}
\examples{
build_ngrams(c("1_1_0", "2_1_0", "5_1_0", "7_1_0", "4_2_0",
"5_2_0", "7_2_0", "8_5_0"))
}
\seealso{
Function used by \code{build_ngrams} to extract unigrams: \code{\link{position_ngrams}}.
}

