% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_cont_table_log_probability}
\alias{calc_cont_table_log_probability}
\title{Calculate the probability of a certain contingency table}
\usage{
calc_cont_table_log_probability(p, q, n11, n01, n10, n00)
}
\arguments{
\item{p}{numeric, a probability of feature equaling 1}

\item{q}{numeric, a probability of feature equaling 1}

\item{n11}{number of elements for which both target and feature equal 1}

\item{n01}{number of elements for which both target and feature equal 1,0}

\item{n10}{number of elements for which both target and feature equal 0,1}

\item{n00}{number of elements for which both target and feature equal 0}
}
\value{
log_probability probability of getting given contingency table
        under p and q
}
\description{
Contingency table for two Bernoulli random variables. Feature and
target.
}
\keyword{internal}

