% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_ed.R
\name{calc_ed}
\alias{calc_ed}
\title{Calculate encoding distance}
\usage{
calc_ed(a, b)
}
\arguments{
\item{a}{list of groups to which elements of sequence should be aggregated.}

\item{b}{list of groups to which \code{a} should be compared.}
}
\value{
an encoding distance.
}
\description{
Compares two encodings and calculates encoding distance between them.
}
\examples{
aa1 = list(`1` = c("g", "a", "p", "v", "m", "l", "i"),
           `2` = c("k", "h"),
           `3` = c("d", "e"),
           `4` = c("f", "r", "w", "y", "s", "t", "c", "n", "q"))

aa2 = list(`1` = c("g", "a", "p", "v", "m", "l", "q"),
           `2` = c("k", "h", "d", "e", "i"),
           `4` = c("f", "r", "w", "y", "s", "t", "c", "n"))
calc_ed(aa1, aa2)
}

