% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_partial_binomial_log_probability}
\alias{calc_partial_binomial_log_probability}
\title{Calculate log of partial binomial probabilities}
\usage{
calc_partial_binomial_log_probability(p, m, n)
}
\arguments{
\item{p}{probability of success}

\item{m}{number of trials}

\item{n}{number of successes}
}
\value{
temp numeric, log(m!/[(m-n)! n!] p^n
}
\description{
Calculate log of partial binomial probabilities
}
\keyword{internal}

