% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_encodings_dist_1.r
\name{change_2_groups}
\alias{change_2_groups}
\title{Generates a list of encodings whose distances from a given encoding are equal to 1.}
\usage{
change_2_groups(enc, i, j, l)
}
\arguments{
\item{enc}{encoding for which all encodings in distance equal to 1 are seached for.}
}
\description{
Generates a list of encodings whose distances from a given encoding are equal to 1.
}

