% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/construct_ngrams.R
\name{construct_ngrams}
\alias{construct_ngrams}
\title{Construct and filter n-grams}
\usage{
construct_ngrams(target, seq, u, n_max)
}
\arguments{
\item{target}{\code{integer} vector with target information (e.g. class labels).}

\item{seq}{\code{integer} vector or matrix describing sequence(s).}

\item{u}{\code{integer}, \code{numeric} or \code{character} vector of all
possible unigrams.}

\item{n_max}{size of constructed n-grams.}
}
\value{
a vector of n-grams.
}
\description{
Step-by-step builds and selects important n-grams.
}
\details{
\code{construct_ngrams} extracts unigrams from the sequences and filters
significant features. From filtered unigrams, the function builds all possible
bigrams using \code{\link{build_ngrams}} and so on, till the required size of
the n-gram is achieved. On each step, obtained n-grams are filtered using
\code{\link{test_features}}.
}
\examples{
deg_seqs <- degenerate(human_cleave[, 1L:9],
list(`1` = c(1, 6, 8, 10, 11, 18),
     `2` = c(2, 13, 14, 16, 17),
     `3` = c(5, 19, 20),
     `4` = c(7, 9, 12, 15),
     '5' = c(3, 4)))
}

