% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_reg_exp.R
\name{count_df}
\alias{count_df}
\title{Compute degrees of freedom for a list of regular expressions}
\usage{
count_df(reg_exp)
}
\arguments{
\item{regExp,}{list of list of vectors, lsit of regular expressions}
}
\value{
numeric, degrees of freedom
}
\description{
Compute degrees of freedom for a list of regular expressions
}
\examples{
x1 <- list(c(1,2),3)
x2 <- list(1,c(2,4))
count_df(list(x1, x2))
}
\keyword{internal}

