% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/count_multigrams.R
\name{count_multigrams}
\alias{count_multigrams}
\title{Detect and count multiple n-grams in sequences}
\usage{
count_multigrams(ns, ds, seq, u, pos = FALSE, scale = FALSE,
  threshold = 0)
}
\arguments{
\item{ns}{\code{numeric} vector of n-grams' sizes. See Details.}

\item{ds}{\code{list} of distances between elements of n-grams. Each element of the list
is a vector used as distance for the respective n-gram size given by the \code{ns}
parameter.}

\item{seq}{\code{integer} vector or matrix describing sequence(s).}

\item{u}{\code{integer}, \code{numeric} or \code{character} vector of all
possible unigrams.}

\item{pos}{\code{logical}, if \code{TRUE} position-specific n_grams are counted.}

\item{scale}{\code{logical}, if \code{TRUE} output data is normalized. Should be
used only for n-grams without position information. See \code{Details}.}

\item{threshold}{\code{integer}, if not equal to 0, data is binarized into
two groups (larger or equal to threshold vs. smaller than threshold).}
}
\value{
a \code{integer} matrix with named columns. The naming conventions are the same
as in \code{\link{count_ngrams}}.
}
\description{
A convinient wrapper around \code{\link{count_ngrams}} for counting multiple
values of \code{n} and \code{d}.
}
\examples{
seqs <- matrix(sample(1L:4, 600, replace = TRUE), ncol = 50)
count_multigrams(c(3, 1), list(c(1, 0), 0), seqs, 1L:4, pos = TRUE)
}

