% Generated by roxygen2 (4.0.1): do not edit by hand
\name{count_ngrams}
\alias{count_ngrams}
\title{Detect and count n-grams in sequences}
\usage{
count_ngrams(seq, n, u, d = 0, pos = FALSE, scale = FALSE,
  threshold = 0)
}
\arguments{
\item{seq}{\code{integer} vector or matrix describing sequence(s).}

\item{n}{\code{integer} size of n-gram.}

\item{u}{unigrams (\code{integer}, \code{numeric} or \code{character} vector).}

\item{d}{\code{integer} vector of distances between elements of n-gram (0 means
consecutive elements). See Details.}

\item{pos}{\code{logical}, if \code{TRUE} n_grams contains position information.}

\item{scale}{\code{logical}, if \code{TRUE} output data is normalized.}

\item{threshold}{\code{integer}, if not equal to 0, data is binarized into
two groups (larger or equal to threshold, smaller than threshold).}
}
\value{
a \code{integer} matrix with named columns. Elements of n-gram are separated
by dot. If \code{pos}, left side of name means actual position of the n-gram
(separated by \code{_}). Right side of name is vector of distance(s) used separated by
\code{_}.
}
\description{
Counts all n-grams present in sequences.
}
\details{
The length of \code{distance} vector should be always \code{n} - 1. For example
when \code{n} = 3, \code{d} = c(1, 2) means A_A__A. For \code{n} = 4,
\code{d} = c(2, 0, 1) means A__AA_A. If vector \code{d} has length 1, it is recycled to
length \code{n} - 1.
}
\note{
List of possible n-grams must be calculated outside of the function.
}
\examples{
#trigrams for nucleotides
count_ngrams(sample(1L:4, 50, replace = TRUE), 3, 1L:4, pos = TRUE)
#trigrams from multiple sequences
seqs <- matrix(sample(1L:4, 600, replace = TRUE), ncol = 50)
count_ngrams(seqs, 3, 1L:4, pos = TRUE)
}
\seealso{
Create vector of possible n-grams: \code{\link{create_ngrams}}.
Get indices of n-grams: \code{\link{get_ngrams_ind}}.
}

