% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_specified.R
\name{count_specified}
\alias{count_specified}
\title{Count specified n-grams}
\usage{
count_specified(seq, ngrams)
}
\arguments{
\item{seq}{\code{integer} vector or matrix describing sequence(s).}

\item{ngrams}{a vector of n-grams. Must have the same \code{n}.}
}
\value{
a \code{\link[slam]{simple_triplet_matrix}} where columns represent
n-grams and rows sequences.
}
\description{
Counts specified n-grams present in the input sequence(s).
}
\examples{
seqs <- matrix(c(1, 2, 2, 1, 2, 1, 2, 1, 1, 2, 3, 4, 1, 2, 2, 4), nrow = 2)
count_specified(seqs, ngrams = c("2_1.1.1_0.0", "2_2.2.2_0.0", "3_1.1.2_0.0"))
}

