% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_feature_target}
\alias{create_feature_target}
\title{Create feature with given contingency matrix}
\usage{
create_feature_target(n11, n01, n10, n00)
}
\arguments{
\item{n11}{number of elements for which both target and feature equal 1}

\item{n01}{number of elements for which both target and feature equal 1,0}

\item{n10}{number of elements for which both target and feature equal 0,1}

\item{n00}{number of elements for which both target and feature equal 0}
}
\value{
a matrix of 2 columns and n11+n10+n01+n00 rows. First column is
target, second is feature
}
\description{
Create feature with given contingency matrix
}
\examples{
temp <- create_feature_target(10, 375, 15, 600)
}

