% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_distribution.R
\docType{class}
\name{criterion_distribution}
\alias{criterion_distribution}
\title{criterion_distribution class}
\description{
A result of \code{\link{distr_crit}} function.
}
\details{
An object of class \code{criterion_distribution} is a numeric matrix.
}
\section{Data}{

\describe{
  \item{1st column:}{possible values of criterion.}
  \item{2nd column:}{probability density function.}
  \item{3rd column:}{cumulative distribution function.}
}
}

\section{Attributes}{

\describe{
  \item{plot_data}{A matrix with values of the criterion and their probabilities.}
  \item{nice_name}{'Nice' name of the criterion.}
}
}

