% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_test_class.R
\name{cut.feature_test}
\alias{cut.feature_test}
\title{Categorize tested features}
\usage{
\method{cut}{feature_test}(x, split = "significances", breaks = c(0, 1e-04, 0.01, 0.05, 1), ...)
}
\arguments{
\item{x}{an object of class \code{\link{feature_test}}.}

\item{split}{attribute along which output should be categorized. Possible values are
\code{"significances"}, \code{"positives"} and \code{"negatives"}. See \code{Value}.}

\item{breaks}{a vector of significances of frequencies along which n-grams are aggregated. 
See description of \code{\link[base]{cut}} function and \code{Details}.}

\item{...}{further parameters accepted by the \code{\link[base]{cut}} function.}
}
\value{
the value of function depends on the \code{split} parameter. 
The function returns a named list of length equal to the length 
of \code{significances} (when \code{split} equals \code{"significances"}) or 
\code{frequencies} (when \code{split} equals \code{"positives"} or \code{"negatives"})
minus one. Each elements of the list contains names of the n-grams belonging to the given 
significance or frequency group.
}
\description{
Categorizes results of \code{\link{test_features}} function into groups based on their 
significance.
}
\keyword{manip}
