% Generated by roxygen2 (4.0.2): do not edit by hand
\name{data_sample}
\alias{data_sample}
\title{Subsample data}
\usage{
data_sample(indices, train_size, test_size)
}
\arguments{
\item{indices}{of data.}

\item{train_size}{size of training sample.}

\item{test_size}{size of test sample.}
}
\value{
a list of a length 2 containing train and test sample indices.
}
\description{
Creates random subsample of population using indices.
}
\details{
Check it for 0 sample size.
}
\examples{
data_sample(1:20, 5, 5)
}

