% Generated by roxygen2 (4.0.2): do not edit by hand
\name{decode_ngrams}
\alias{decode_ngrams}
\title{Decode n-grams}
\usage{
decode_ngrams(ngrams)
}
\arguments{
\item{ngrams}{a vector of n-grams.}
}
\value{
a list of length equal to the number of positions present in n-grams. Each element of
the list contains unigrams that are present on this position.
}
\description{
Tranforms a vector of positioned n-grams into a list of positions filled with unigrams
appearing on them.
}

