% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degenerate.R
\name{degenerate_ngrams}
\alias{degenerate_ngrams}
\title{Degenerate n-grams}
\usage{
degenerate_ngrams(x, element_groups, binarize = FALSE)
}
\arguments{
\item{x}{a \code{\link[slam]{simple_triplet_matrix}} or matrix of n-grams.}

\item{element_groups}{encoding of elements: list of groups to which elements 
of n-grams should be aggregated. Must have unique names.}

\item{binarize}{logical indicating if n-grams should be binarized}
}
\value{
Depending on the \code{x{}} a \code{\link[slam]{simple_triplet_matrix}} 
or matrix of degenerated n-grams.
}
\description{
'Degenerates' n-grams by aggregating amino acid or nucleotide elements
into bigger groups.
}
