% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degenerate.R
\name{full2simple}
\alias{full2simple}
\title{Convert encoding from full to simple format}
\usage{
full2simple(x)
}
\arguments{
\item{x}{encoding.}
}
\description{
Converts an encoding from the full format to the simple format.
}
\examples{
aa1 = list(`1` = c("g", "a", "p", "v", "m", "l", "i"), 
           `2` = c("k", "h"), 
           `3` = c("d", "e"), 
           `4` = c("f", "r", "w", "y", "s", "t", "c", "n", "q"))
full2simple(aa1)

}
