% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngrams.R
\name{gap_ngrams}
\alias{gap_ngrams}
\title{Gap N-Grams}
\usage{
gap_ngrams(ngrams)
}
\arguments{
\item{ngrams}{a vector of positioned n-grams (as created by \code{\link{count_ngrams}}).}
}
\value{
A \code{character} vector of (n-1)-grams ith introduced n-grams.
}
\description{
Introduces gaps in the n-grams.
}
\examples{
gap_ngrams(c("2_1.1.2_0.1", "3_1.1.2_0.0", "3_2.2.2_0.0"))
}

