% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequence.R
\name{generate_sequence}
\alias{generate_sequence}
\title{Generate sequence}
\usage{
generate_sequence(alphabet, regions)
}
\arguments{
\item{alphabet}{the unigram alphabet. Columns are equivalent to unigrams 
and rows to particular properties.}

\item{regions}{a list of rules describing regions.}
}
\description{
Generate a sequences using an alphabet of unigrams and set of rules.
}
