% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequence.R
\name{generate_single_unigram}
\alias{generate_single_unigram}
\title{Generate single unigram}
\usage{
generate_single_unigram(unigram_ranges)
}
\arguments{
\item{unigram_ranges}{list of ranges containing respective properties. If named, 
names are preserved.}
}
\description{
Assign randomly generated properties to a single unigram.
}
\examples{
generate_single_unigram(list(P1 = c(0, 0.5), 
                             P2 = c(0.2, 0.4),
                             P3 = c(0.5, 1),
                             P4 = c(0, 0)))

}
\seealso{
\code{generate_single_unigram} is a helper function for 
\code{\link{generate_unigrams}}.
}
