% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_ngrams_ind}
\alias{get_ngrams_ind}
\title{Get indices of n-grams}
\usage{
get_ngrams_ind(len_seq, n, d)
}
\arguments{
\item{len_seq}{\code{integer} value describing sequence's length.}

\item{n}{\code{integer} vector of n-grams' sizes.}

\item{d}{distance between elements of n-gram (0 means consecutive elements). See
Details.}
}
\value{
A list with number of elements equal to \code{n}. Every element is a
vector containing locations of given n-gram letter. For example, first element of
list contain indices of first letter of all n-grams. The attribute \code{d} of output
contains distances between letter used to compute locations (see Details).
}
\description{
Computes list of n-gram elements positions in sequence.
}
\details{
The length of \code{d} vector should be always n - 1. For example
when \code{n} = 3, \code{d} = c(1, 2) means A_A__A. For \code{n} = 4,
\code{d} = c(2, 0, 1) means A__AA_A. Shorter \code{d} vectors will be recycled.
}
\examples{
#trigrams in sequence of length 10
get_ngrams_ind(10, 9, 0)
}

