% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/criterion_distribution.R
\name{ggplot2.criterion_distribution}
\alias{ggplot2.criterion_distribution}
\title{Plot criterion distribution with ggplot2}
\usage{
ggplot2.criterion_distribution(x)
}
\arguments{
\item{x}{object of class \code{\link{criterion_distribution}}.}
}
\value{
ggplot object.
}
\description{
Plots results of \code{\link{distr_crit}} function.with package ggplot2
}

