% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gramize_data}
\alias{gramize_data}
\title{Build n-gram table}
\usage{
gramize_data(dat, d2, d3, scale)
}
\arguments{
\item{dat}{sequence table.}

\item{d2}{distances for 2-grams.}

\item{d3}{distances for 3-grams.}

\item{scale}{\code{logical}, if \code{TRUE} output data is normalized.}
}
\value{
a dataframe of all n-grams counts.
}
\description{
Builds table containing counts of 1-, 2- and 3- grams from numeric sequence table.
}
\details{
Input data must be a matrix or data frame of numeric elements.
}

