% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{human_cleave}
\alias{human_cleave}
\title{Human signal peptides cleavage sites}
\format{A data frame with 1296 observations on the following 10 variables.
\describe{
\item{P1}{amino acid in first position.}
\item{P9}{amino acid in nineth position.}
\item{tar}{target vector (1 if cleavage site, 0 if post-cleavage site).}}}
\description{
uniprot query \code{keyword:signal AND taxonomy:"Homo sapiens (Human) [9606]"
AND annotation:(type:signal AND evidence:ECO_0000269) AND reviewed:yes}
cleaves[i] + 5):(cleaves[i] + 13)]
}
\examples{
data(human_cleave)
table(human_cleave[, 1])
}
\keyword{datasets}

