% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ig_distribution}
\alias{ig_distribution}
\title{Computes Information Gain distribution for feature, target under null hypothesis}
\usage{
ig_distribution(target, feature, graphical.output = FALSE)
}
\arguments{
\item{target}{{0,1}-valued target vector}

\item{feature}{{0,1}-valued feature vector}

\item{graphical.output}{default value is FALSE}
}
\value{
a matrix of 3 rows. First row are possible IG values,
 second in probability density function, third is cimulative
 distribution function
}
\description{
Computes Information Gain distribution for feature, target under null hypothesis
}
\examples{
temp <- create_feature_target(10, 375, 15, 600)
ig_distribution(temp[,1], temp[,2], graphical.output = TRUE)
}

