% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_ngram.R
\name{is_ngram}
\alias{is_ngram}
\title{Validate n-gram}
\usage{
is_ngram(x)
}
\arguments{
\item{x}{a \code{character} string representing single n-gram.}
}
\value{
\code{TRUE} if n-gram's notation is correct, \code{FALSE} if not.
}
\description{
Checks if the character string may be used as an n-gram and its notation follows specific
convention of \code{biogram} package.
}
\examples{
print(is_ngram("1_1.1.1_0.0"))
print(is_ngram("not_ngram"))
}

