% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_reg_exp.R
\name{join_reg_exp}
\alias{join_reg_exp}
\title{Join regular expressions}
\usage{
join_reg_exp(exp1, exp2)
}
\arguments{
\item{e1}{list of vectors, first regular expression}

\item{e2}{list of vectors, second regular expression}
}
\value{
list of vectors, samllest regular expression that contains e1 and e2
}
\description{
Join regular expressions
}
\examples{
x1 <- list(c(1,2),3)
x2 <- list(1,c(2,4))
join_reg_exp(x1, x2)
}
\keyword{internal}

