% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengths_ngrams.R
\name{lengths_ngrams}
\alias{lengths_ngrams}
\title{Get lengths of the n-grams}
\usage{
lengths_ngrams(ngrams)
}
\arguments{
\item{ngrams}{a \code{character} vector of n-grams.}
}
\value{
A \code{numeric} vector of n-gram lengths.
}
\description{
Computes the length of n-grams.
}
\examples{
lengths_ngrams(c("2_1.1.2_0.1", "3_1.1.2_2.0", "3_2.2.2_0.0"))
}
