% Generated by roxygen2 (4.0.2): do not edit by hand
\name{letters2numbers}
\alias{letters2numbers}
\title{Convert letters to numbers}
\usage{
letters2numbers(seq, what = "rna")
}
\arguments{
\item{seq}{\code{character} vector representing single sequence.}

\item{what}{the type of sequence. Can be \code{rna}, \code{dna} or \code{prot}.}
}
\value{
a \code{numeric} vector containing converted elements.
}
\description{
Converts biological sequence of letters to number notation.
}
\examples{
sample_seq <- c("a", "d", "d", "g", "a", "g", "n", "a", "l")
letters2numbers(sample_seq, "prot")
}
\keyword{manip}

