% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Matrix2matrix}
\alias{Matrix2matrix}
\title{Sparse Matrix to Numeric Matrix}
\usage{
Matrix2matrix(object)
}
\arguments{
\item{object}{\code{\link[Matrix]{dgCMatrix-class}} or
\code{\link[Matrix]{lgCMatrix-class}} matrix, usually a result of the
\code{\link{count_ngrams}} function.}
}
\value{
A \code{numeric} matrix of n-grams.
}
\description{
Converts objects of \code{\link[Matrix]{dgCMatrix-class}} or
\code{\link[Matrix]{lgCMatrix-class}} to numeric matrix.
}
\examples{
x1 <- count_ngrams(sample(1L:4, 50, replace = TRUE), 3, 1L:4, pos = FALSE)
class(x1)
x2 <- Matrix2matrix(x1)
class(x2)
}

