% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/criterion_distribution.R
\name{plot.criterion_distribution}
\alias{plot.criterion_distribution}
\title{Plot criterion distribution}
\usage{
\method{plot}{criterion_distribution}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link{criterion_distribution}}.}

\item{...}{further arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
nothing.
}
\description{
Plots results of \code{\link{distr_crit}} function.
}

