% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for further analysis}
\usage{
prepare_data(data_set, pos, neg, train_size, test_size)
}
\arguments{
\item{data_set}{data set.}

\item{pos}{indices of positive data.}

\item{neg}{indices of negative data.}

\item{train_size}{a vector of length 2 - size of positive and negative training
subsample.}

\item{test_size}{a vector of length 2 - size of positive and negative test
subsample.}
}
\value{
a list of a length 2 containing cases belonging to respectively train and test
sample.
}
\description{
Creates random stratified subsample of population.
}
\details{
Check it for 0 sample size.
}
\examples{
#data(iris)
#prepare_data(iris, 1:20, 101:120, c(5, 5), c(8, 8))
}

