% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Read FASTA files}
\usage{
read_fasta(file)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}
}
\value{
a list of sequences.
}
\description{
A lightweight tool to read nucleic or amino-acid sequences from a file 
in FASTA format.
}
\examples{
\dontrun{
  read_fasta("https://www.uniprot.org/uniprot/P28307.fasta")
  }
}
\seealso{
\code{\link[seqinr]{read.fasta}}: heavier function for processing FASTA files.
}
