% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_txt}
\alias{read_txt}
\title{Read sequences from .txt file}
\usage{
read_txt(connection)
}
\arguments{
\item{connection}{a \code{\link{connection}} to the text (.txt) file.}
}
\value{
a list of sequences.
}
\description{
Read sequence data saved in text file.
}
\details{
The input file should contain one or more amino acid sequences separated by 
empty line(s).
}
\examples{
sequences <- read_txt(system.file("PlastoGram/sequences.txt", package = "PlastoGram"))
}
\keyword{manip}
