% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_reg_exp.R
\name{reg_exp_df}
\alias{reg_exp_df}
\title{Compute degrees of freedom of one regular expression}
\usage{
reg_exp_df(x)
}
\arguments{
\item{x,}{list of vectors, regular expression}
}
\value{
numeric, degrees of freedom
}
\description{
Compute degrees of freedom of one regular expression
}
\keyword{internal}

