% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\docType{class}
\name{regional_param}
\alias{regional_param}
\title{regional_param class}
\description{
List of rules defining the region.
}
\details{
An object of the \code{regional_param} class is a list consisting of all rules 
necessary to properly build a region.
}
\section{Attributes}{

\describe{
  \item{reg_len}{the number of unigrams inside the region. Might be 0}
  \item{prop_ranges}{required intervals of properties of unigrams in the region}
  \item{exactness}{a \code{numeric} value between 0 and 1 defining how stricly 
unigrams are kept within \code{prop_ranges}. If 1, only unigrams within 
\code{prop_ranges} are inside the region. if 0.9, there is 10% chance that 
unigrams that are not in the \code{prop_ranges} will be inside the region.}
}
}

\seealso{
\code{\link{generate_sequence}}
}
