% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degenerate.R
\name{simple2full}
\alias{simple2full}
\title{Convert encoding from simple to full format}
\usage{
simple2full(x)
}
\arguments{
\item{x}{encoding (see Details).}
}
\description{
Converts an encoding from the simple format to the full format.
}
\details{
The encoding should be named. Each name should correspond to a different
amino acid or nucleotide.
}
\examples{
aa1 = structure(c("1", "4", "3", "3", "4", "1", "2", "1", "2", "1", 
                  "1", "4", "1", "4", "4", "4", "4", "1", "4", "4"), 
                .Names = c("a", "c", "d", "e", "f", "g", "h", "i", 
                           "k", "l", "m", "n", "p", "q", 
                           "r", "s", "t", "v", "w", "y"))
simple2full(aa1)

}
