% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify_alph}
\alias{simplify_alph}
\title{Simplify alphabets}
\usage{
simplify_alph(target, seqs, n_generations = 100, n_alph = 200,
  alph_len = 6)
}
\arguments{
\item{target}{\code{integer} vector with target information (e.g. class labels).}

\item{n_generations}{number of generations.}

\item{n_alph}{number of alphabets in a generation.}

\item{alph_len}{length of the alphabet.}
}
\value{
A list.
}
\description{
'Degenerates' amino acid or nucleic sequence by aggregating 
elements to bigger groups.
}
\keyword{manip}
