% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.feature_test}
\alias{summary.feature_test}
\title{Summarize tested features}
\usage{
\method{summary}{feature_test}(object, significance = 0.05, ...)
}
\arguments{
\item{object}{of class \code{\link{feature_test}}.}

\item{significance}{a threshold level. A feature with p-value equal or below it is considered
significant.}

\item{...}{ignored}
}
\value{
nothing.
}
\description{
Summarizes results of \code{\link{test_features}} function.
}
\keyword{manip}
\keyword{methods}
\keyword{print}

