% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabulate_ngrams.R
\name{table_ngrams}
\alias{table_ngrams}
\title{Tabulate n-grams}
\usage{
table_ngrams(seq, ngrams, target)
}
\arguments{
\item{seq}{\code{integer} vector or matrix describing sequence(s).}

\item{ngrams}{a vector of n-grams. Must have the same \code{n}.}

\item{\code{integer}}{vector with target information (e.g. class labels).}
}
\value{
a data frame with the number of columns equal to the length of the
\code{target} plus 1. The first column contains names of the n-grams. Further
columns represents frequency of n-grams for respective value of the
\code{target}.
}
\description{
Builds a contingency table of the n-gram counts versus their class labels.
}

