% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_ig}
\alias{test_ig}
\title{Test information gains}
\usage{
test_ig(target, features, times)
}
\arguments{
\item{target}{target vector.}

\item{features}{\code{integer} matrix of features with number of rows equal
to the length of target vector.}

\item{times}{number of times procedure should be repetead}
}
\value{
a numeric vector of lenth equal to the number of features containing computed
information gain values.
}
\description{
Computes p-value of features' information gain.
}
\note{
Both \code{target} and \code{features} must be binary, i.e. contain only 0
and 1 values.
}
\examples{
calc_ig(sample(0L:1, 100, replace = TRUE),
matrix(sample(0L:1, 400, replace = TRUE), ncol = 4))
}
\seealso{
\code{\link{calc_ig}}
}

