% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_enc.R
\name{write_encoding}
\alias{write_encoding}
\title{Write encodings to a file}
\usage{
write_encoding(x, file = "")
}
\arguments{
\item{x}{encoding or list of encodings.}

\item{file}{ither a character string naming a file or a 
\code{\link[base]{connection}} 
open for writing. "" indicates output to the console.}
}
\description{
Saves a list of encodings (or a single encoding to the file).
}
\examples{
aa1 = list(`1` = c("g", "a", "p", "v", "m", "l", "i"), 
           `2` = c("k", "h"), 
           `3` = c("d", "e"), 
           `4` = c("f", "r", "w", "y", "s", "t", "c", "n", "q"))
write_encoding(aa1)

}
