% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Write FASTA files}
\usage{
write_fasta(seq, file, nchar = 80)
}
\arguments{
\item{seq}{a list of sequences.}

\item{file}{the name of the output file.}

\item{nchar}{the number of characters per line.}
}
\description{
A lightweight tool to read nucleic or amino-acid sequences from a file 
in FASTA format.
}
\seealso{
\code{\link[seqinr]{write.fasta}}: heavier function for writing FASTA files.
}
