% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aggregate.feature_test}
\alias{aggregate.feature_test}
\title{Aggregate tested features}
\usage{
\method{aggregate}{feature_test}(x, significances = c(0, 1e-04, 0.01, 0.05,
  1), ...)
}
\arguments{
\item{x}{an object of class \code{\link{feature_test}}.}

\item{significances}{a vector of significances along which p-values are classified. See description of
\code{\link[base]{cut}} function.}

\item{...}{ignored}
}
\value{
a named list of length equal to the length of \code{significances} minus one. Each elements of
the list contains names of the n-grams belonging to the given significance group.
}
\description{
Aggregates results of \code{\link{test_features}} function into groups based on their
significance.
}
\keyword{manip}

