% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ed.R
\name{calc_ed}
\alias{calc_ed}
\title{Calculate encoding distance}
\usage{
calc_ed(a, b)
}
\arguments{
\item{a}{an encoding (see Note)}

\item{b}{an encoding to which \code{a} should be compared. Must be have less groups than 
\code{a} or have equal length.}
}
\value{
an encoding distance.
}
\description{
Computes encoding distance between two encodings.
}
\details{
The encoding distance between \code{a} and \code{b} is defined as the minimum number 
of amino acids that have to be moved between subgroups of encoding to make \code{a} identical to
\code{b} (order of subgroups in the encoding and amino acids in a group is unimportant).
}
\note{
The encoding is a list of groups to which elements of sequence should be aggregated.
}
\examples{
#calculate encoding distance between two encodings of amino acids
aa1 = list(`1` = c("g", "a", "p", "v", "m", "l", "i"), 
           `2` = c("k", "h"), 
           `3` = c("d", "e"), 
           `4` = c("f", "r", "w", "y", "s", "t", "c", "n", "q"))

aa2 = list(`1` = c("g", "a", "p", "v", "m", "l", "q"), 
           `2` = c("k", "h", "d", "e", "i"), 
           `4` = c("f", "r", "w", "y", "s", "t", "c", "n"))
calc_ed(aa1, aa2) 
    
#the encoding distance between two identical encodings is 0
calc_ed(aa1, aa1) 
 
}

