% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_coding.R
\name{code_ngrams}
\alias{code_ngrams}
\title{Code n-grams}
\usage{
code_ngrams(decoded_ngrams)
}
\arguments{
\item{decoded_ngrams}{a \code{character} vector of decoded n-grams.}
}
\value{
a \code{character} vector of n-grams.
}
\description{
Code human-friendly representation of n-grams into a biogram format.
}
\examples{
code_ngrams(c("11_2", "1__12", "222"))
code_ngrams(c("aaa_b", "d__aa", "abd"))
}
\seealso{
Inverse function: \code{\link{decode_ngrams}}.
}

