% Generated by roxygen2 (4.0.1): do not edit by hand
\name{count_ngrams}
\alias{count_ngrams}
\title{Detect and count n-grams in single sequence}
\usage{
count_ngrams(seq, feature_list, n, d = 0, scale = FALSE)
}
\arguments{
\item{seq}{\code{integer} vector describing sequence.}

\item{feature_list}{list of features (see Note).}

\item{n}{size of n-grams.}

\item{d}{distance between elements of n-gram (0 means consecutive elements). See
Details.}

\item{scale}{\code{logical}, if \code{TRUE} output data is normalized.}
}
\value{
a named \code{integer} vector. Elements of n-gram are separated by dot.
}
\description{
Counts all n-grams present in given sequence.
}
\details{
The length of \code{distance} vector should be always n - 1. For example
when \code{n} = 3, \code{d} = c(1, 2) means A_A__A. For \code{n} = 4,
\code{d} = c(2, 0, 1) means A__AA_A.
}
\note{
List of possible n-grams must be calculated outside of the function.
}
\examples{
#trigrams for nucleotides
count_ngrams(sample(1L:4, 30, replace = TRUE), create_ngrams(3, 1L:4), 3)
}
\seealso{
For convenient wrapper for multidistance calculations
see \code{\link{mdcount_ngrams}}.
}

