% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_ngrams}
\alias{create_ngrams}
\title{Get all possible n-grams}
\usage{
create_ngrams(n, u)
}
\arguments{
\item{n}{size of gram (i.e. 2-grams: "11", "12", ...).}

\item{u}{unigrams.}
}
\value{
a character vector. Elements of n-gram are separated by dot.
}
\description{
Creates vector of all posible n_grams.
}
\details{
Input data must be a matrix or data frame of numeric elements.
}
\examples{
#bigrams for standard aminoacids
create_ngrams(2, 1L:20)
}

