% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/decode_ngrams.R
\name{decode_ngrams}
\alias{decode_ngrams}
\title{Decode n-grams}
\usage{
decode_ngrams(ngrams)
}
\arguments{
\item{ngrams}{a vector of n-grams.}
}
\value{
a \code{character} vector of length equal to the number of n-grams.
}
\description{
Tranforms a vector of n-grams into a more human-friendl form.
}
\examples{
decode_ngrams(c("2_1.1.2_0.0", "3_1.1.2_0.0", "3_2.2.2_0.0"))
}
\seealso{
Other n-gram deciphering functions: \code{\link{position_ngrams}}.
}

