% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram_coding.R
\name{decode_ngrams}
\alias{decode_ngrams}
\title{Decode n-grams}
\usage{
decode_ngrams(ngrams)
}
\arguments{
\item{ngrams}{a \code{character} vector of n-grams.}
}
\value{
a \code{character} vector of length equal to the number of n-grams.
}
\description{
Transforms a vector of n-grams into a human-friendly form.
}
\note{
Decoded n-grams lose the position information.
}
\examples{
decode_ngrams(c("2_1.1.2_0.1", "3_1.1.2_2.0", "3_2.2.2_0.0"))
}
\seealso{
Validate n-gram structure: \code{\link{is_ngram}}.

Inverse function: \code{\link{code_ngrams}}.
}

