% Generated by roxygen2 (4.0.2): do not edit by hand
\name{degenerate}
\alias{degenerate}
\title{Degenerate protein sequence}
\usage{
degenerate(seq, aa_group)
}
\arguments{
\item{seq}{\code{character} vector representing single aminoacid sequence.}

\item{aa_group}{list of aminoacid groups to which sequence should be aggregated.}
}
\value{
a \code{character} vector.
}
\description{
'Degenerates' protein sequence by aggregating aminoacids to bigger groups.
}
\examples{
sample_seq <- c(1, 3, 1, 3, 4, 4, 3, 1, 2)
table(sample_seq)

#aggregate sequence to purins and pyrimidines
deg_seq <- degenerate(sample_seq, list(W = c(1, 4), S = c(2, 3)))
table(deg_seq)
}
\keyword{manip}

