% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_test_class.R
\docType{class}
\name{feature_test}
\alias{feature_test}
\title{feature_test class}
\description{
A result of \code{\link{test_features}} function.
}
\details{
An object of class \code{feature_test} is a numeric vector of p-values. Additional 
attributes characterizes futher the details of test which returned these p-values.
}
\section{Attributes}{

\describe{
  \item{criterion}{the criterion used in permutation test.}
  \item{adjust}{the name of p-value adjusting method}
  \item{times}{the number of permutations.  If QuiPT was chosen \code{NA}.}
  \item{occ}{frequency of features splitted in subset based on the value of target.}
}
}
\seealso{
Methods: 
\itemize{
\item \code{\link{as.data.frame.feature_test}}
\item \code{\link{cut.feature_test}}
\item \code{\link{print.feature_test}}
\item \code{\link{summary.feature_test}}
}
}

