% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/feature_test_class.R
\docType{class}
\name{feature_test}
\alias{feature_test}
\title{feature_test class}
\description{
A result of \code{\link{test_features}} function.
}
\details{
An object of class \code{test_features} is a numeric vector of p-values. Additional
attributes characterizes futher the details of test which returned these p-values.
}
\section{Attributes}{

\describe{
  \item{criterion}{the criterion used in permutation test.}
  \item{adjust}{the name of p-value adjusting method}
  \item{times}{the number of permutations.  If QuiPT was chosen \code{NA}.}
}
}

