% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ig_distribution}
\alias{ig_distribution}
\title{Computes Information Gain distribution for feature, target under null hypothesis}
\usage{
ig_distribution(target, feature, graphical.output = FALSE, criterion = "ig")
}
\arguments{
\item{target}{{0,1}-valued target vector}

\item{feature}{{0,1}-valued feature vector}

\item{graphical.output}{default value is FALSE, if TRUE
probability density function is plotted}

\item{criterion}{the criterion used for calculations of distribution.}
}
\value{
a matrix of 3 rows. First row are possible IG values,
 second in probability density function, third is cimulative
 distribution function
}
\description{
Computes Information Gain distribution for feature, target under null hypothesis
}
\examples{
target_feature <- create_feature_target(10, 375, 15, 600)
ig_distribution(target = target_feature[,1], feature = target_feature[,2],
graphical.output = TRUE)
}
\seealso{
\code{\link{calc_ig}}
}

