% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/degenerate.R
\name{l2n}
\alias{l2n}
\title{Convert letters to numbers}
\usage{
l2n(seq, seq_type)
}
\arguments{
\item{seq}{\code{character} vector representing single sequence.}

\item{seq_type}{the type of sequence. Can be \code{rna}, \code{dna} or \code{prot}.}
}
\value{
a \code{numeric} vector containing converted elements.
}
\description{
Converts biological sequence from letter to number notation.
}
\examples{
sample_seq <- c("a", "d", "d", "g", "a", "g", "n", "a", "l")
l2n(sample_seq, "prot")
}
\seealso{
\code{l2n} is a wrapper around \code{\link{degenerate}}.

Inverse function: \code{\link{n2l}}.
}
\keyword{manip}

