% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2matrix.R, R/utilities.R
\name{list2matrix}
\alias{list2matrix}
\title{Convert list of sequences to matrix}
\usage{
list2matrix(seq_list)

list2matrix(seq_list)
}
\arguments{
\item{seq_list}{\code{list} of sequences, e.g. \code{\link[seqinr]{SeqFastaAA}}.}

\item{seq_list}{list of sequences (e.g. as returned by 
the \code{\link[seqinr]{read.fasta}} function).}
}
\value{
A matrix with the number of rows equal to the number of sequences and the 
number of columns equal to the length of the longest sequence.

a matrix of sequences.
}
\description{
Converts list of sequences to matrix.

Converts a list of sequences to a matrix.
}
\details{
a matrix with the number of rows equal to the number of 
sequences and the number of columns equal to the length of the 
longest sequence. If sequences do not have the same length, 
shorter sequences are filled with NAs.
}
\note{
Since matrix must have specified number of columns, ends of shorter sequences are 
completed with NAs.
}
\examples{
list2matrix(list(s1 = c("c", "g", "g", "t"),
                 s2 = c("g", "t", "c", "t", "t", "g"),
                 s3 = c("a", "a", "t")))
# list of sequences
sample_list <- lapply(round(runif(5, 6, 15), 0), function(i)
  sample(c("A", "C", "G", "T"), i, replace = TRUE))
# convert to matrix
list2matrix(sample_list)
}
\keyword{manip}

