% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/degenerate.R
\name{n2l}
\alias{n2l}
\title{Convert numbers to letters}
\usage{
n2l(seq, seq_type)
}
\arguments{
\item{seq}{\code{numeric} vector representing single sequence.}

\item{seq_type}{the type of sequence. Can be \code{rna}, \code{dna} or \code{prot}.}
}
\value{
a \code{numeric} vector containing converted elements.
}
\description{
Converts biological sequence from number to letter notation.
}
\examples{
sample_seq <- c(1, 3, 3, 6, 1, 6, 12, 1, 10)
n2l(sample_seq, "prot")
}
\seealso{
\code{n2l} is a wrapper around \code{\link{degenerate}}.

Inverse function: \code{\link{l2n}}.
}
\keyword{manip}

