% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/position_ngrams.R
\name{position_ngrams}
\alias{position_ngrams}
\title{Position n-grams}
\usage{
position_ngrams(ngrams)
}
\arguments{
\item{ngrams}{a vector of positioned n-grams (as created by \code{\link{count_ngrams}}).}
}
\value{
a list of length equal to the number of unique positions present in n-grams. Each
element of the list contains unigrams that are present on this position.
}
\description{
Tranforms a vector of positioned n-grams into a list of positions filled with unigrams
appearing on them.
}
\examples{
position_ngrams(c("2_1.1.2_0.1", "3_1.1.2_0.0", "3_2.2.2_0.0"))
}
\seealso{
Change n-gram name to human-friendly form: \code{\link{decode_ngrams}}.

Validate n-gram structure: \code{\link{is_ngram}}.
}

