% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_test_class.R
\name{print.feature_test}
\alias{print.feature_test}
\title{Print tested features}
\usage{
\method{print}{feature_test}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link{feature_test}}.}

\item{...}{further arguments passed to \code{\link[base]{print.default}}.}
}
\value{
nothing.
}
\description{
Prints results of \code{\link{test_features}} function.
}
