% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_test_class.R
\name{summary.feature_test}
\alias{summary.feature_test}
\title{Summarize tested features}
\usage{
\method{summary}{feature_test}(object, conf_level = 0.95, ...)
}
\arguments{
\item{object}{of class \code{\link{feature_test}}.}

\item{conf_level}{confidence level. A feature with p-value equal to or smaller than the 
confidence is considered significant.}

\item{...}{ignored}
}
\value{
nothing.
}
\description{
Summarizes results of \code{\link{test_features}} function.
}
\keyword{manip}
