% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test_features_fast}
\alias{test_features_fast}
\title{Permutation test for feature selection}
\usage{
test_features_fast(target, features, criterion = "ig")
}
\arguments{
\item{target}{target vector.}

\item{features}{\code{integer} matrix of features with number of rows equal
to the length of target vector.}

\item{criterion}{the criterion used in permutation test.}
}
\value{
a numeric vector of lenth equal to the number of features containing computed
information gain values.
}
\description{
Performs a feature selection on positioned N-gram data using a Fisher's
permutation test.
}
\details{
Currently implemented criterions:
\itemize{
\item{"\code{ig}" - information gain}
}
}
\note{
Both \code{target} and \code{features} must be binary, i.e. contain only 0
and 1 values.
}
\examples{
temp <- create_feature_target(10, 390, 0, 600)
temp2 <- create_feature_target(9, 391, 1, 599)
temp3 <- create_feature_target(8, 392, 0, 600)
test_features_fast(temp[,1], cbind(temp[,2], temp2[,2], temp3[,2]))
}
\seealso{
\code{\link{calc_ig}}, \code{\link{ig_distribution}}
}

